refactor: reorganize Discord integration classes into distinct packages for better structure and add corresponding schema, doc too
