+++ src/main/java/org/phong/zenflow/plugin/subdomain/nodes/builtin/integration/discord/trigger/DiscordMessageTriggerExecutor.java
@@ -1,10 +1,11 @@
-package org.phong.zenflow.plugin.subdomain.nodes.builtin.integration.discord;
+package org.phong.zenflow.plugin.subdomain.nodes.builtin.integration.discord.trigger;
 
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import net.dv8tion.jda.api.JDA;
 import org.phong.zenflow.plugin.subdomain.execution.dto.ExecutionResult;
 import org.phong.zenflow.plugin.subdomain.node.registry.PluginNode;
+import org.phong.zenflow.plugin.subdomain.nodes.builtin.integration.discord.share.DiscordJdaResourceManager;
 import org.phong.zenflow.workflow.subdomain.context.ExecutionContext;
 import org.phong.zenflow.workflow.subdomain.node_definition.definitions.dto.WorkflowConfig;
 import org.phong.zenflow.workflow.subdomain.trigger.infrastructure.persistence.entity.WorkflowTrigger;
