+++ src/main/java/org/phong/zenflow/plugin/subdomain/nodes/builtin/integration/discord/trigger/schema.json
@@ -1,0 +1,152 @@
+{
+  "$schema": "http://json-schema.org/draft-07/schema#",
+  "type": "object",
+  "properties": {
+    "input": {
+      "type": "object",
+      "properties": {
+        "bot_token": {
+          "type": "string",
+          "description": "Discord bot token for authentication. Keep this secure and never expose it in logs.",
+          "minLength": 50
+        },
+        "channel_id": {
+          "type": "string",
+          "description": "Discord channel ID to monitor for messages. Must be a valid Discord snowflake ID.",
+          "pattern": "^[0-9]{17,19}$"
+        },
+        "content_contains": {
+          "type": "string",
+          "description": "Optional filter - only trigger if message content contains this text."
+        },
+        "ignore_bots": {
+          "type": "boolean",
+          "default": true,
+          "description": "Whether to ignore messages from bot accounts."
+        },
+        "user_filter": {
+          "type": "array",
+          "items": {
+            "type": "string",
+            "pattern": "^[0-9]{17,19}$"
+          },
+          "description": "Optional array of user IDs to monitor. If specified, only messages from these users will trigger workflows."
+        },
+        "role_filter": {
+          "type": "array",
+          "items": {
+            "type": "string",
+            "pattern": "^[0-9]{17,19}$"
+          },
+          "description": "Optional array of role IDs. If specified, only users with these roles can trigger workflows."
+        },
+        "command_prefix": {
+          "type": "string",
+          "description": "Optional command prefix (e.g., '!' or '/'). Only messages starting with this prefix will trigger workflows."
+        },
+        "case_sensitive": {
+          "type": "boolean",
+          "default": false,
+          "description": "Whether content filtering should be case-sensitive."
+        }
+      },
+      "required": ["bot_token", "channel_id"],
+      "additionalProperties": true
+    },
+    "output": {
+      "type": "object",
+      "properties": {
+        "trigger_type": {
+          "type": "string",
+          "const": "discord_message"
+        },
+        "triggered_at": {
+          "type": "number",
+          "description": "Unix timestamp when the trigger fired"
+        },
+        "message_id": {
+          "type": "string",
+          "description": "Discord message ID (snowflake)"
+        },
+        "channel_id": {
+          "type": "string",
+          "description": "Discord channel ID where the message was sent"
+        },
+        "author_id": {
+          "type": "string",
+          "description": "Discord user ID of the message author"
+        },
+        "content": {
+          "type": "string",
+          "description": "Full text content of the Discord message"
+        },
+        "guild_id": {
+          "type": "string",
+          "description": "Discord server (guild) ID"
+        },
+        "timestamp": {
+          "type": "string",
+          "format": "date-time",
+          "description": "ISO timestamp when the message was created"
+        },
+        "author": {
+          "type": "object",
+          "description": "Author information",
+          "properties": {
+            "id": {"type": "string"},
+            "username": {"type": "string"},
+            "discriminator": {"type": "string"},
+            "avatar": {"type": "string"},
+            "is_bot": {"type": "boolean"}
+          }
+        },
+        "channel": {
+          "type": "object",
+          "description": "Channel information",
+          "properties": {
+            "id": {"type": "string"},
+            "name": {"type": "string"},
+            "type": {"type": "string"}
+          }
+        },
+        "guild": {
+          "type": "object",
+          "description": "Server (guild) information",
+          "properties": {
+            "id": {"type": "string"},
+            "name": {"type": "string"}
+          }
+        },
+        "attachments": {
+          "type": "array",
+          "description": "Array of message attachments",
+          "items": {
+            "type": "object",
+            "properties": {
+              "id": {"type": "string"},
+              "filename": {"type": "string"},
+              "url": {"type": "string"},
+              "size": {"type": "number"}
+            }
+          }
+        },
+        "mentions": {
+          "type": "object",
+          "description": "Mentions in the message",
+          "properties": {
+            "users": {
+              "type": "array",
+              "items": {"type": "string"}
+            },
+            "roles": {
+              "type": "array",
+              "items": {"type": "string"}
+            },
+            "everyone": {"type": "boolean"}
+          }
+        }
+      }
+    }
+  }
+}
+
